fix flushing
