--- agent/agent-tooling/src/main/resources/META-INF/services/io.opentelemetry.sdk.autoconfigure.spi.internal.AutoConfigureListener
@@ -1,1 +1,0 @@
-com.microsoft.applicationinsights.agent.internal.init.FlushOnShutdown
